Make Everson Mono default font for better Unicode.
Everson Mono TTF can be found for all major platforms, and it covers a large
portion of BMP, meaning Emacs doesn't have to use fallback fonts which ruin the
"mono" aspect of mono-spaced.  This is not usually important, but in settings
where whitespace is significant (e.g. Markdown Tables), it really helps to have
good visual feedback.

In order to play well with font-lock-mode, the system default monospace font
should also be set to Everson Mono.