pandoc include tag: recognize multipl tagged regions
This extends the my/pandoc-include-tag function to search for multiple matched
pairs of the tag given in some tag spec, and returns all the delineated regions
joined together.  This currently ensures at least one newline between the text
of the regions, but it might be worth expanding the tag format to specify a
separator to use.  Not sure if that's too much.