add `man` keybinding for emacs
This is also the first commit where I'm putting closing parens on their own line
with justification unrelated to aesthetics. The rationale is the same as that of
trailing commas for list-like structures in some languages, namely, it's
diff-friendly. If the list-closing paren is alone, additions to the list show up
as just that: additions.  If not, there will be a subtraction for the line of
the previous trailing element which now has fewer close parens.

This should also apply to any sexpr in lisp-land that accepts an arbitrary
number of elements (i.e. anything `progn`-like), spans multiple lines, *and* is
likely to change.