WIP: attempting to pause music on lock/suspend
I don't think this can actually be done easily with just i3 and execing.
More likely i3exit should be modified a bit, and a systemd unit should be made
to hook into the resume.  It would be nice if i3lock had some kind of
customizable hook action, but that still wouldn't solve suspend/hibernate.