fix linuxbrew prefix
location of `brew` is used to set homebrew variables. The intent
with linuxbrew is that the prefix is /home/linuxbrew/.linuxbrew
and you use $prefix/bin/brew. Thats usually a symlink, and
using the target is a mistake (messes up the prefix, which in
turn messes up installation of Bottles)